Move tests and implementation of each Funlet into separate subfolders
This is consistent with the new structure of the project.
