+++ funlet-find-me/tests/test-whisper.sh
