+++ funlet-simple-menu/tests/test-simple-menu.sh
