+++ funlet-simple-message/functions/funlet-simple-message.js
