+++ funlet-simple-message/tests/test-simple-message.sh
