+++ funlet-simulring/functions/funlet-simulring.js
