+++ funlet-simulring/tests/test-whisper.sh
