+++ funlet-whisper/functions/funlet-whisper.js
