+++ funlet-whisper/tests/test-whisper.sh
